requirements.txt
https://download.pytorch.org/whl/cu126/torch-2.6.0%2Bcu126-cp310-cp310-win_amd64.whl#sha256=eda7768f0a2ad9da3513abf60ff5c13049e7e2ec74ed4cfcd4736a8523ab1f89
https://download.pytorch.org/whl/cu126/torchvision-0.21.0%2Bcu126-cp310-cp310-win_amd64.whl#sha256=d6b23af252e8f4fc923d57efeab5aad7a33b6e15a72a119d576aa48ec1e0d924
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
fake-useragent==2.0.3
filelock==3.13.1
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.2.10
fsspec==2024.6.1
humanfriendly==10.0
idna==3.10
imageio==2.31.2
imread_from_url==0.1.3
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.3.2
lazy_loader==0.3
libauc==1.2.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
onnx==1.17.0
onnxruntime-gpu==1.19.2
opencv-python==4.10.0.84
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
pyreadline3==3.5.4
qudida==0.0.4
requests==2.32.3
scikit-image==0.21.0
sympy==1.13.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
